**Issue**
Resolves #10044

This commit makes some of the tests faster by avoiding using StorageService unless neccessary.

In`ert/unit_tests/gui/simulation/test_run_dialog.py` Tests affected:
* `test_that_design_matrix_show_parameters_button_is_visible`: 10.5s -> 1.01s
* `test_that_exception_in_base_run_model_is_handles`: 4.71s -> 1.04s

In`ert/ui_tests/gui/test_main_window.py`
Tests affected:
* `test_that_es_mda_restart_run_box_is_disabled_when_there_are_no_cases`: 4.17s -> 1.57s
* `test_help_menu`: 4.15s -> 0.53s

In`ert/ui_tests/gui/test_restart_no_responses_and_parameters.py` Tests affected:
* `test_sensitivity_restart`: 4.53s -> 1.83s

In`ert/ui_tests/gui/test_workflow_tool.py`
Tests affected:
* `test_run_export_runpath_workflow`: 6.54s -> 4.57s

In`ert/ui_tests/gui/simulation/test_run_path_dialog.py` Tests affected:
* `test_run_path_deleted_error`: 14.98s -> 7.86s
* `test_run_path_is_deleted`: 8.77s -> 6.39s
* `test_run_path_is_not_deleted`: 8.35s -> 6.33s

Results while testing `pytest tests/ert -n logical`:
this commit: 467.71s
main:  486.76s
(Screenshot of new behavior in GUI if applicable)


- [x] PR title captures the intent of the changes, and is fitting for release notes.
- [x] Added appropriate release note label
- [ ] Commit history is consistent and clean, in line with the [contribution guidelines](https://github.com/equinor/ert/blob/main/CONTRIBUTING.md).
- [x] Make sure unit tests pass locally after every commit (`git rebase -i main
      --exec 'just rapid-tests'`)

## When applicable
- [ ] **When there are user facing changes**: Updated documentation
- [ ] **New behavior or changes to existing untested code**: Ensured that unit tests are added (See [Ground Rules](https://github.com/equinor/ert/blob/main/CONTRIBUTING.md#ground-rules)).
- [ ] **Large PR**: Prepare changes in small commits for more convenient review
- [ ] **Bug fix**: Add regression test for the bug
- [ ] **Bug fix**: Create Backport PR to latest release

<!--
Adding labels helps the maintainers when writing release notes. This is the [list of release note labels](https://github.com/equinor/ert/labels?q=release-notes).
-->