@jonathan-eq 2025-02-20
Comparing with https://github.com/equinor/ert/actions/runs/13433700710/job/37531087794:
gui-test py3.11: 10m40s -> 8m38s
gui-test py3.12: 12m27s -> 10m44s
performance-test py3.11 6m52s -> 6m27s
performance-test py3.12 8m5s-> 7m57s
cli-test py3.11: 13m3s ->13m4s
cli-test py3.12: 15m43s -> 15m59s